Fix importing JSX into tests
